deploy.sh: Switch to per-release dockerhub repos
The idea now is that there will be one repo for each release and master,
along with the floating repo that tracks the current release. So
curently there will be toaster-krogoth, toaster-master and toaster
repos.

The toaster repo will match the toaster-krogoth repo, but is provided so
that new users can just do a "docker pull toaster" and know they are
using the latest release.

Signed-off-by: Randy Witt <randy.e.witt@linux.intel.com>