Added test to make sure the container priming works
The toaster container is primed, meaning that the database is already
set up and the layers are downloaded.  This greatly increases the speed
with which the container starts up.  This adds tests to make sure that
the migrations are not being run and that the layers are not being
downloaded again.

Signed-off-by: bavery <brian.avery@intel.com>