tests/runtests.sh: Add LOCALDIR environment variable
If you have LOCALDIR set in your environment then when you execute
runtests.sh it uses LOCALDIR for the toaster workdir and runs the
container with --local set.  This requires that LOCALDIR has a poky
directory in it.  The primary use case for this is so that you can run
the tests against a local checkout of poky that you are actively
developing.

Signed-off-by: bavery <brian.avery@intel.com>