sudoers.usersetup: keep env variables passed to docker run
Environment variables passed to docker run via -e such as -e
"http_proxy=myproxy.com:8080" should be available to the user running
toaster.  This adds SETENV to the sudoers.usersetup file so that
usersetup.py has the right to call sudo with a --preserve-env flag.
This is to keep in sync with the usersetup.py which comes from the
extensible sdk container.

Signed-off-by: bavery <brian.avery@intel.com>