Dockerfile: added comments for Workaround.
We have a Django version workaround in the Dockerfile, I added comments
indicating which releases this workaround is needed for and which
releases the workaround is not needed for. Once we are no longer
supporting the point releases that do not have the fix in the code, we
can retire the workaround.

Signed-off-by: brian avery <brian.avery@intel.com>