.travis.yml: Don't deploy, and remove secure fields
Since the build on travis hasn't been successful for months, take this
opportunity to remove the deploy step. In actuality, a travis image
hasn't been deployed with the failures anyway.

Also remove the secure variables that were used by travis, those can be
replaced if the build is ever re-enabled on travis.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>