workflows,deploy: reinstate FLOATING_REPO
In the Travis CI workflow, we used to track the "LATEST_RELEASE_REPO" as
the latest stable release and that was pushed as the crops/toaster
"FLOATING_REPO".

With the recent release of 'kirkstone', set the LATEST_RELEASE_REPO to
match.

While we are at it, reinstate the date-commitish tags that were also
part of the Travis CI workflow.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>